fix: clang -Wconversion errors and SonarCloud findings
CI caught two clang -Werror failures that gcc does not diagnose, so every
local run was green:

  spanned_zip.cpp read_u16_le  the shift promotes uint16_t to int, so the
                               `|` result is an int narrowed on return
  archive_test_helpers.h       same shape in the uudecode helper
  spanned_zip.cpp              unused PK_LOCAL_HEADER (the merger never
                               validates local headers)

All three were in code harvested from subagents, and clangd had flagged the
conversions inline — I dismissed them as stale. They were not. Ran the full
clang parity build locally this time rather than round-tripping through CI.

SonarCloud findings, all 35:

- archive_kind: magic signatures move from string literals to byte arrays.
  `\x` is greedy in C++, so a magic byte followed by a hex-digit character
  silently absorbs it — these are binary anyway, and the array says so.
  (Kills five S7040 at once rather than escaping around them.)
- Two cognitive-complexity refactors, both split along seams that already
  existed rather than shuffled to satisfy the metric:
    walk_one 38 -> apply_own_meta / place_media / vet_nested / descend
    merge_spanned_zip 42 -> rewrite_central_directory extracted, which also
      carried both nested breaks (S924)
- Mechanical: init-statements, `using enum`, `auto` for redundant types,
  const& for RecursionLimits, const Frame&, empty() over size()==0,
  emplace_back, explicit & on function pointers, merged nested ifs, and a
  while loop whose counter is not its condition.

The spanned-ZIP end-to-end test (real `zip -s` set, merged, reopened through
miniz) still passes, which is what proves the extraction preserved
behaviour rather than just compiling.

1438 tests, 0 failed under scripts/test.sh, --asan and --tsan; gcc AND
clang builds clean under -Werror.

+++ tests/ui/archive_test_helpers.h
@@ -157,12 +157,15 @@ inline std::vector<uint8_t> uudecode(const fs::path& uufile, const fs::path& out
             const uint8_t c2 = (line[pos++] - 32) & 0x3F;
             const uint8_t c3 = (line[pos++] - 32) & 0x3F;
 
-            decoded.push_back((c0 << 2) | (c1 >> 4));
+            // Each operand promotes to int through << / |, so the narrowing
+            // back to a byte must be explicit — gcc accepts it, clang's
+            // -Wconversion does not.
+            decoded.push_back(static_cast<uint8_t>((c0 << 2) | (c1 >> 4)));
             if (i + 1 < line_len) {
-                decoded.push_back((c1 << 4) | (c2 >> 2));
+                decoded.push_back(static_cast<uint8_t>((c1 << 4) | (c2 >> 2)));
             }
             if (i + 2 < line_len) {
-                decoded.push_back((c2 << 6) | c3);
+                decoded.push_back(static_cast<uint8_t>((c2 << 6) | c3));
             }
         }
     }
